Refactor: Update Navigation and UI Components
This commit introduces several changes related to navigation strategies and UI component styling.

- **Navigation Strategy Refactor:**
    - Moved `DialogStrategy` and `TwoPaneNav3` (renamed to `TwoPaneSceneStrategy`) into a new `strategy` subpackage within `com.programmersbox.uiviews.presentation.navigation`. This improves the organization of navigation-related code.
    - Updated imports in `NavGraphView` and `Nav3Graph` to reflect the new package structure for these strategies.
    - `Nav3Graph`:
        - Commented out an experimental `animatedEntry` function that aimed to provide `LocalNavigationAnimatedScope`.

- **MoreSettingsScreen UI:**
    - Added icons to various preference settings in `MoreSettingsScreen` for better visual cues:
        - Export/Import Favorites: `Publish` and `Download` icons.
        - Sync Cloud/Local Favorites: `Sync` and `CloudSync` icons.
        - Export All/Specific Lists: `Publish` icon.
        - Import List: `Download` icon.

- **GeneralSettings UI:**
    - `GeneralSettings.kt` (kmpuiviews): Added `padding(4.dp)` to the `FlowRow` containing theme color selection for better spacing.
    - `GeneralSettings.kt` (UIViews): Set `colors = FloatingToolbarDefaults.vibrantFloatingToolbarColors()` for the `HorizontalFloatingToolbar` for a more vibrant appearance.

- **SelectableMiniPalette UI:**
    - Changed the `color` of the `Surface` in `SelectableMiniPalette` from `MaterialTheme.colorScheme.inverseOnSurface` to `MaterialTheme.colorScheme.surfaceContainerHigh`.
