+++ UIViews/src/main/java/com/programmersbox/uiviews/presentation/settings/GeneralSettings.kt
@@ -18,6 +18,7 @@ import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
 import androidx.compose.material3.FilledIconButton
+import androidx.compose.material3.FloatingToolbarDefaults
 import androidx.compose.material3.HorizontalFloatingToolbar
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
@@ -134,6 +135,7 @@ private fun MultipleActionsSetting(
         ) {
             HorizontalFloatingToolbar(
                 expanded = true,
+                colors = FloatingToolbarDefaults.vibrantFloatingToolbarColors(),
                 leadingContent = {
                     var showMenu by remember { mutableStateOf(false) }
                     DropdownMenu(
